Revert "moved mocap to ekf , and extracted setpoint providers"
This reverts commit f8a1bc96bea301ae1616d59c46f7d706058aba8e.
